Separate out autocomplete logic
